requirements.txt
wheel
twine
numpy~=1.23.5
datapane~=0.15.5
matplotlib~=3.6.2
pandas~=1.5.2
altair~=4.2.0
scikit-learn~=1.2.0
tqdm~=4.64.1
sklearn-utils
seaborn~=0.12.1
folktables~=0.0.11
xgboost~=1.7.2
munch~=2.5.0